optimizer module split
